Make the remote api functions return api models.
Move the mapping from the api models to core models to the domain usecases.
