+++ slackhealthbot/domain/fitbit/usecase_get_last_sleep.py
@@ -1,4 +1,5 @@
 import datetime
+import logging
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
@@ -16,8 +17,39 @@ async def do(
         db,
         fitbit_userid=fitbit_userid,
     )
-    last_sleep: SleepData = await sleepapi.get_sleep(
+    last_sleep: sleepapi.FitbitSleep = await sleepapi.get_sleep(
         oauth_token=user.oauth_data,
         when=when,
     )
-    return last_sleep
+    return parse_sleep(last_sleep) if last_sleep else None
+
+
+def parse_sleep(fitbit_sleep: sleepapi.FitbitSleep) -> SleepData | None:
+    main_sleep_item = next(
+        (item for item in fitbit_sleep.sleep if item.isMainSleep), None
+    )
+    if not main_sleep_item:
+        logging.warning("No main sleep found")
+        return None
+
+    wake_minutes = (
+        main_sleep_item.levels.summary.awake.minutes
+        if main_sleep_item.type == "classic"
+        else main_sleep_item.levels.summary.wake.minutes
+    )
+    asleep_minutes = (
+        main_sleep_item.levels.summary.asleep.minutes
+        if main_sleep_item.type == "classic"
+        else main_sleep_item.duration / 60000 - wake_minutes
+    )
+    return SleepData(
+        start_time=datetime.datetime.strptime(
+            main_sleep_item.startTime, DATETIME_FORMAT
+        ),
+        end_time=datetime.datetime.strptime(main_sleep_item.endTime, DATETIME_FORMAT),
+        sleep_minutes=asleep_minutes,
+        wake_minutes=wake_minutes,
+    )
+
+
+DATETIME_FORMAT = "%Y-%m-%dT%H:%M:%S.%f"
